requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.26.123
botocore==1.29.123
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.4.0
Django==4.1.4
django-celery-beat==2.5.0
django-ckeditor==6.5.1
django-environ==0.10.0
django-js-asset==2.0.0
django-storages==1.13.2
django-timezone-field==5.0
hiredis==2.2.3
idna==3.4
jmespath==1.0.1
kavenegar==1.1.2
kombu==5.2.4
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
